Move badges to the bottom of README.rst
Just a suggestion. I don't like seeing lots and lots of badges at the top of the page. The only really important badges for the end user IMHO are docs (for obvious reasons) and pypi version (to direct the user to pypi if encountered elsewhere), though even pypi could be removed from the top since it's in the install instructions anyway. The rest can safely be moved to the bottom, for easy access whilst avoiding the "in your face" feeling you get from the gargantuan badge table when it's at the top.