Update heading in README.rst
Since README.rst is included in docs/index.rst, the heading will be shown in the ReadTheDocs menu. I don't think the project name is especially informative as a title in the menu, so I'm using "Overview" here in my packages. The package name is shown at the top on PyPI anyway.

Just a suggestion. :)