Add option ci_https_proxy and use it to define .gitlab-ci.yml so that GitLab can run its CI pipeline.
Delete .gitlab-ci.yml if we are not using GitLab.

If we put null as the default, then for some reason that argument is required instead of defaulting to null.

If the user sets an SSH_PRIVATE_KEY, then set up SSH.

Use prebuilt dahanna/python.3.7-git-tox-alpine with git and tox already installed.

Deploy GitLab Pages even if tests fail, because if we include a link to the documentation in the README, then sphinx linkcheck will fail, so the pages will never get deployed, so sphinx linkcheck will keep failing.

Use python -m sphinx, so that we can preinstall sphinx and use it with --sitepackages.