Add GNU LGPL licenses.
This commit adds support for GNU LGPL v3 (or later), or v2 (or later)
to the license options.

Note that GNU GPLv3 is an add-on to the GNU GPLv3, so the LICENSE file
includes both licenses in this case.

GNU licensing guidelines (see: https://www.gnu.org/licenses/gpl-howto.html)
recommend placing a copyright and license notice at the top of every
file; to keep consistency with how the other licenses are implemented
in `cookiecutter-pylibrary`, we put such a notice at the top of the
`LICENSE` file only.